% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/AR.R
\name{AR.DArch}
\alias{AR.DArch}
\title{Calculates the Accruacy Ratio of a given set of probability}
\usage{
\method{AR}{DArch}(x, input = x@dataSet@data, target = x@dataSet@targets,
  ...)
}
\arguments{
\item{x}{a DArch instance}

\item{input}{the input matrix}

\item{target}{binary response}

\item{...}{additional inputs}
}
\description{
This function calculates the Accuracy Ratio of a trained darch instance
}

