% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/util.R
\name{AR.default}
\alias{AR.default}
\title{Utility function that calculates the Accuracy Ratio of a binary classification
 model}
\usage{
\method{AR}{default}(x, target)
}
\arguments{
\item{x}{a list of probability}

\item{target}{target variable}
}
\description{
Calculates the Accruacy Ratio of a given set of probability
}

