% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/AR.R
\name{AR}
\alias{AR}
\title{Calculates the Accuracy Ratio of a classifier}
\usage{
AR(x, ...)
}
\arguments{
\item{x}{model}

\item{...}{additional inputs}
}
\description{
This function calculates the Accuracy Ratio of a binary classification
 model
}

