% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/backprop_delta.R
\name{backpropagate_delta}
\alias{backpropagate_delta}
\title{function that calculates the delta function of a darch object}
\usage{
backpropagate_delta(darch, trainData, targetData, errorFunc = meanSquareErr)
}
\description{
Calculates the delta functions using backpropagation
}

