% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/batch_normalization.R
\name{batch_normalization}
\alias{batch_normalization}
\title{Batch Normalization Function that normalizes the input before applying non-linearity}
\usage{
batch_normalization(x, gamma, beta, mu = NULL, sigma_2 = NULL,
  epsilon = exp(-12))
}
\arguments{
\item{x}{weighted sum of outputs from the previous layer}

\item{gamma}{the gamma coefficient}

\item{beta}{the beta coefficient}

\item{epsilon}{a constant added to the variance for numerical stability}
}
\description{
Normalize the input distribution for hidden layers in a neural network
}
\references{
http://jmlr.org/proceedings/papers/v37/ioffe15.pdf Pg 4
}

