% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/batch_normalization.R
\name{batch_normalization_differential}
\alias{batch_normalization_differential}
\title{Function that calcualtes the differentials in the batch normalization mode}
\usage{
batch_normalization_differential(delta_y, mu, sigma_2, x, x_hat, y, gamma, beta)
}
\arguments{
\item{delta_y}{derivative wrt y}

\item{mu}{mean of the input}

\item{sigma_2}{variance of the input}

\item{x}{input}

\item{x_hat}{normalized input}

\item{y}{transformed input after batch normalization}

\item{gamma}{gamma coefficient}

\item{beta}{beta coefficient}
}
\description{
Calculates the differentials in batch normalization
}
\references{
http://jmlr.org/proceedings/papers/v37/ioffe15.pdf Pg 4
}

