% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/calculate_mu_sigma.R
\name{calcualte_population_mu_sigma}
\alias{calcualte_population_mu_sigma}
\title{Calculates the mu and sigmas of a darch instance}
\usage{
calcualte_population_mu_sigma(darch, input)
}
\arguments{
\item{darch}{a darch instance}

\item{input}{input data}
}
\description{
This function calculates the mu and sigmas of hidden layers in a darch instance
}

