% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/error_functions.R
\name{classification_error}
\alias{classification_error}
\title{Calculates the classification error}
\usage{
classification_error(output, target)
}
\arguments{
\item{output}{the output of a classifier in the form of probability. Probability > 1
will be treated as positive (target = 1).}

\item{target}{the target variable}
}
\description{
This function calculates the classification error
}

