% Generated by roxygen2 (4.0.1): do not edit by hand
\name{Comparison}
\alias{!=}
\alias{!=.Layer}
\alias{==}
\alias{==.Layer}
\alias{Comparison}
\title{Layers equality testing}
\usage{
\method{==}{Layer}(e1, e2)

\method{!=}{Layer}(e1, e2)
}
\arguments{
\item{e1,e2}{Layers to compare}
}
\description{
Binary operator to test for equality or inequality of two \code{Layer}s. This is equal to a call to \code{\link{identical}}.
}
\examples{
Layer(10L, "gaussian") == Layer(10L, "gaussian") # TRUE
Layer(10L, "gaussian") == Layer(10L, "continuous") # FALSE
Layer(10L, "gaussian") == Layer(20L, "gaussian") # FALSE
Layer(10L, "gaussian") != Layer(10L, "gaussian") # FALSE
Layer(10L, "gaussian") != Layer(10L, "continuous") # TRUE
Layer(10L, "gaussian") != Layer(20L, "gaussian") # TRUE
}
\seealso{
\code{\link{Comparison}} \code{\link{Layer}}
}

