% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/continue.functions.R
\name{continue.functions}
\alias{continue.function.always}
\alias{continue.function.exponential}
\alias{continue.function.exponential.aic}
\alias{continue.function.exponential.bic}
\alias{continue.function.random}
\alias{continue.functions}
\title{Continue functions}
\usage{
continue.function.exponential(error, iter, batchsize, maxiters, layer = 0,
  ic = AIC)

continue.function.always(error, iter, batchsize, maxiters, layer = 0)

continue.function.exponential.aic(error, iter, batchsize, maxiters, layer = 0)

continue.function.exponential.bic(error, iter, batchsize, maxiters, layer = 0)

continue.function.random(error, iter, batchsize, maxiters, layer = 0)
}
\arguments{
\item{error}{a vector of the errors along the training}

\item{iter, batchsize}{current iteration number and batchsize.}

\item{maxiters}{maximum number of iterations}

\item{layer}{during RBM pre-training, which layer is being pre-trained. Otherwise, 0.}

\item{ic}{the Information Criterion to use, typically \code{\link{AIC}} or \code{\link{BIC}}}
}
\value{
boolean (see description)
}
\description{
Functions to stop (by returning \code{FALSE}) the training if it has converged.

\code{continue.function.exponential} fits an exponential to the error and return \code{TRUE} if the function hasn't converged (or in case of doubt), \code{FALSE} if the function has converged or if the data couldn't be fitted (plateau reached or no exponential fit).
As a side effect, this function plots the error and the fit, and prints a summary of the fit on the console

\code{continue.function.always} always returns true so that the training carries on until \code{maxiters} is reached
}

