% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/preprocess_data.R
\name{convert_categorical}
\alias{convert_categorical}
\title{Data proprosess function that covnerts a categorical input to continuous input or
vectorize it}
\usage{
convert_categorical(x, type = "ordinal", ordered_list = list(),
  var_name = "var", ...)
}
\arguments{
\item{x}{input variable}

\item{type}{ordinal or other}

\item{ordered_list}{the rank ordering of an ordinal variable. Users are expected to
provide a complete list of the rank ordering. Otherwise, a default rank ordering
will be used.}

\item{var_name}{the name of the input variable. This is used to to create vectorized
input variables}

\item{...}{other inputs}
}
\description{
Proprosess a data set. It converts categorical data into binary variables
if it is unordered or continuous variable from 0 to 1 if it is ordinal
}

