% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/error_functions.R
\name{crossEntropyErr}
\alias{crossEntropyErr}
\title{Calculates the cross entropy error}
\usage{
crossEntropyErr(output, target)
}
\arguments{
\item{output}{the output value}

\item{target}{the target value}
}
\description{
This function calculates the cross entropy error and its first order derivatives
}

