% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/generics.R
\name{drop}
\alias{drop}
\title{Drop a DeepBeliefNet}
\usage{
drop(x)
}
\arguments{
\item{x}{the DeepBeliefNet}
}
\description{
Drops the DeepBeliefNet information if it contains only a single RestrictedBolzmannMachine.
Provides an alias to \code{\link[base]{drop}} as \code{drop.default}.
}
\examples{
data(pretrained.mnist)
dbn <- DeepBeliefNet(Layer(784, "continuous"), Layer(1000, "binary"))
print(dbn)
rbm <- drop(dbn)
print(rbm)
}
\seealso{
\code{\link{DeepBeliefNet}}, \code{\link[base]{drop}}
}

