% Generated by roxygen2 (4.0.0): do not edit by hand
\name{error}
\alias{error}
\alias{error.DeepBeliefNet}
\alias{error.RestrictedBolzmannMachine}
\alias{rmse}
\alias{rmse.DeepBeliefNet}
\alias{rmse.RestrictedBolzmannMachine}
\title{Compute Reconstruction Error}
\usage{
error(...)

rmse(...)

\method{error}{DeepBeliefNet}(x, data, ...)

\method{rmse}{DeepBeliefNet}(x, data, ...)

\method{error}{RestrictedBolzmannMachine}(x, data, ...)

\method{rmse}{RestrictedBolzmannMachine}(x, data, ...)
}
\arguments{
\item{x}{the \code{\link{RestrictedBolzmannMachine}} or \code{\link{DeepBeliefNet}} object}

\item{data}{feature vector}

\item{...}{further arguments to the \code{plot} function above and to the \code{predict} function.}
}
\value{
the reconstruction error of the data
}
\description{
Computes the reconstruction error (rmse) of the prediction of the data. \code{rmse} is an alias for \code{error}.
}

