% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/evidence.gradient.R
\name{evidence.gradient}
\alias{evidence.gradient}
\alias{evidence.gradient.RestrictedBolzmannMachine}
\title{Computes and returns the gradient of the evidence function}
\usage{
evidence.gradient(...)

\method{evidence.gradient}{RestrictedBolzmannMachine}(x, data, ...)
}
\arguments{
\item{...}{ignored}

\item{x}{the \code{\link{RestrictedBolzmannMachine}} object}

\item{data}{feature vector}
}
\value{
the reconstruction error of the data
}
\description{
This is specifically the contrastive divergence evidence.
}

