% Generated by roxygen2 (4.0.0): do not edit by hand
\name{Extract}
\alias{Extract}
\alias{[}
\alias{[.DeepBeliefNet}
\alias{[[}
\alias{[[.DeepBeliefNet}
\alias{[[<-}
\alias{[[<-.DeepBeliefNet}
\title{Extract or Replace Parts of a Deep Belief Net}
\usage{
\method{[[}{DeepBeliefNet}(x, i)

\method{[[}{DeepBeliefNet} (x, i) <- value

\method{[}{DeepBeliefNet}(x, i, drop = FALSE)
}
\arguments{
\item{x}{the DBN}

\item{i}{indices specifying elements to extract or replace}

\item{value}{the RestrictedBolzmannMachine to insert.}

\item{drop}{whether to drop the DeepBeliefNet if it contains a single RestrictedBolzmannMachine.}
}
\description{
Operators to extract or replace parts of a \code{\link{DeepBeliefNet}} object.
In extraction methods, the \code{pretrained} and \code{finetuned} switches will match those of the DeepBeliefNet that was supplied, while \code{unrolled} will be set to \code{FALSE}.
In replacement methods, the \code{pretrained} switch will be on if all the \code{RestrictedBolzmannMachines} are pretrained, while \code{finetuned} and \code{unrolled} will be set to \code{FALSE}.
}
\details{
\code{[[} extracts (and \code{[[<-} replaces) \emph{exactly} one \code{\link{RestrictedBolzmannMachine}} layer of the DeepBeliefNet.

\code{[} extracts one or more layers of the DeepBeliefNet. If the returned DeepBeliefNet. has exactly one RestrictedBolzmannMachine, the \code{drop}
argument controls whether the function returns an \code{\link{RestrictedBolzmannMachine}} object (\code{TRUE})
or a \code{\link{DeepBeliefNet}} object containing one single \code{\link{RestrictedBolzmannMachine}} (\code{FALSE}).
}
\section{Note}{

If the DeepBeliefNet contains N layers, there are N-1 RestrictedBolzmannMachines.
}
\examples{
dbn <- DeepBeliefNet(Layers(c(784, 1000, 500, 250, 30), input="continuous", output="gaussian"))
# Extract a RBM
dbn[[2]]
# Replace a RBM
dbn[[1]] <- RestrictedBolzmannMachine(Layer(10, "binary"), Layer(1000, "binary"))
dbn[[2]] <- RestrictedBolzmannMachine(Layer(1000, "binary"), Layer(500, "binary"))
dbn[[4]] <- RestrictedBolzmannMachine(Layer(250, "binary"), Layer(2, "gaussian"))
\dontrun{
# Cannot replace incompatible RestrictedBolzmannMachines
dbn[[2]] <- RestrictedBolzmannMachine(1000, 400, input="binary", output="binary")
dbn[[2]] <- RestrictedBolzmannMachine(100, 500, input="binary", output="binary")
dbn[[2]] <- RestrictedBolzmannMachine(1000, 500, input="binary", output="continuous")
dbn[[2]] <- RestrictedBolzmannMachine(1000, 500, input="gaussian", output="binary")
}
# Get the first layer as RestrictedBolzmannMachine
dbn[[1]]
dbn[1, drop=TRUE]
# Get the first layer as DeepBeliefNet
dbn[1]
}
\seealso{
\code{\link{DeepBeliefNet}}, \code{\link{RestrictedBolzmannMachine}}
}

