% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/finetune_SGD.R
\name{finetune_SGD}
\alias{finetune_SGD}
\title{Finetune function that trains a deep neural network using stochastic gradient descent method}
\usage{
finetune_SGD(darch, trainData, targetData, errorFunc = meanSquareErr)
}
\description{
This function finetunes a DArch network using SGD approach
}

