% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/finetune_SGD_bn.R
\name{finetune_SGD_bn}
\alias{finetune_SGD_bn}
\title{Finetune function that trains a deep neural network using stochastic gradient descent method}
\usage{
finetune_SGD_bn(darch, trainData, targetData, errorFunc = meanSquareErr)
}
\description{
This function finetunes a DArch network using SGD approach
}

