% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dropout.R
\name{generateDropoutMasksForDarch}
\alias{generateDropoutMasksForDarch}
\title{Generates dropout masks for dnn}
\usage{
generateDropoutMasksForDarch(darch, dropout_input, dropout_hidden)
}
\arguments{
\item{darch, }{a DArch instance}

\item{dropout_input, }{the dropout rate for the input layer}

\item{dropout_hidden, }{the dropout rate for the hidden layer}
}
\description{
This function generates dropout maks for dnn
}
\references{
Dropout: A Simple Way to Prevent Neural Networks from
 Overfitting, Nitish Srivastava
}
\seealso{
\url{https://www.cs.toronto.edu/~hinton/absps/JMLRdropout.pdf}
}

