% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Layer.R
\name{Layer}
\alias{Layer}
\alias{Layers}
\title{Layer of a DeepBeliefNet}
\usage{
Layer(size, type)

Layers(sizes, input, output, hidden = "binary")
}
\arguments{
\item{size, sizes}{the number(s) of nodes of the Layer(s). Will be truncated to integers.}

\item{type, input, output, hidden}{the type of the Layer, or for Layer the types of input, output and hidden layers. If omitted, \code{hidden}
layers will be implicitly assigned a \dQuote{binary} type. \code{input} and \code{output} layer types must be specified explicitly.}
}
\value{
an object of class \code{layer} containing the following elements:
\itemize{
\item{size}{number of nodes}
\item{type}{types of the nodes. Valid types are \dQuote{binary}, \dQuote{continuous} or \dQuote{gaussian}}
}
}
\description{
This class represents a layer of the DeepBeliefNet.

\code{Layer} creates and returns a single layer of class \code{layer}, while \code{Layers} returns a list of \code{Layer}s of the given sizes, input and output. Hidden (or internal) layers are of type \dQuote{binary} by default.
}
\examples{
Layer(10L, "gaussian")
Layers(c(10L, 5L, 2L), "gaussian", "gaussian")
}
