% Generated by roxygen2 (4.0.0): do not edit by hand
\name{length}
\alias{length}
\alias{length.DeepBeliefNet}
\title{Length (or depth) of a DeepBeliefNet}
\usage{
\method{length}{DeepBeliefNet}(x)
}
\arguments{
\item{x}{DeepBeliefNet object}
}
\description{
Computes the number of layers of the DeepBeliefNet. Note that this is not the same
as the number of RestrictedBolzmannMachine contained in the DeepBeliefNet (which is length - 1)
}
\examples{
dbn <- DeepBeliefNet(Layers(c(784, 1000, 500, 250, 30), input="continuous", output="gaussian"))
length(dbn)
}

