% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/util.R
\name{matMult}
\alias{matMult}
\title{Calculates the outer product of two matricies}
\usage{
matMult(data, weight)
}
\arguments{
\item{data}{the date matrix}

\item{weight}{the weight matrix}
}
\description{
Calcualtes the outer product of two matrices
}

