% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/error_functions.R
\name{meanSquareErr}
\alias{meanSquareErr}
\title{Calculates the mean squared error}
\usage{
meanSquareErr(output, target)
}
\arguments{
\item{output}{the output value}

\item{target}{the target value}
}
\description{
This function calculates the mean squared error and its first order derivatives
}

