% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mnist-data.R
\docType{data}
\name{mnist}
\alias{mnist}
\alias{pretrained.mnist}
\alias{trained.mnist}
\title{Pre-trained and fine-tuned DeepBeliefNets for the MNIST dataset}
\description{
These DBNs have already been pre-trained and fine-tuned to model the MNIST dataset. 
They can be used to avoid long training steps, especially in examples of the package documentation.
}
\keyword{data}

