% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/new_dnn.R
\name{new_dnn}
\alias{new_dnn}
\title{Creats a new instance of darch class}
\usage{
new_dnn(layer_structure, layer_functions = NULL,
  output_layer_default = linearUnitDerivative,
  hidden_layer_default = rectified_linear_unit_function,
  weight_initiliazaiton = generateWeights)
}
\arguments{
\item{layer_structure}{a int vector that specifies the number and width of layers}

\item{layer_functions}{a list of activation functions used by each layer}

\item{weight_initiliazaiton}{function that initialize a layer's weight matrix}

\item{layer_function_default}{default function used as the activation function}
}
\description{
This function creates a new instance of darch class
}

