% Generated by roxygen2 (4.0.1): do not edit by hand
\name{predict}
\alias{predict}
\alias{predict.DeepBeliefNet}
\alias{predict.RestrictedBolzmannMachine}
\title{Predict Methods for Deep Belief Nets and Restricted Bolzman Machines}
\usage{
\method{predict}{DeepBeliefNet}(object, newdata, drop = TRUE, ...)

\method{predict}{RestrictedBolzmannMachine}(object, newdata, drop = TRUE, ...)
}
\arguments{
\item{object}{the model}

\item{newdata}{a \code{\link{data.frame}} or \code{\link{matrix}} providing the data. Must have the same columns than the input layer of the model.}

\item{drop}{do not return additional dimensions}

\item{\dots}{ignored}
}
\description{
Obtain predictions from a \code{\link{DeepBeliefNet}} or \code{\link{RestrictedBolzmannMachine}} object
}
\examples{
data(mnist)
## Make predictions on a DBN object
dbn <- DeepBeliefNet(Layers(c(784, 1000, 500, 250, 30), input="continuous", output="gaussian"))
pretrained <- pretrain(dbn, mnist$train$x,
                       penalization = "l2", lambda=0.0002, epsilon=c(.1, .1, .1, .001),
                       batchsize = 100, maxiters=100000)
predict(pretrained, mnist$test$x)
## Make predictions on a RBM object
rbm <- RestrictedBolzmannMachine(Nv = 784, Nh = 1000, input="continuous", output="binary")
pretrained <- pretrain(rbm, mnist$train$x)
predict(pretrained, mnist$test$x)
}

