% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/util.R
\name{print_layer}
\alias{print_layer}
\title{Utility function that prints out the layer value for a neural net}
\usage{
print_layer(darch, num_of_layer, input = darch@dataSet@data,
  show_derivative = F, type = "heatmap")
}
\arguments{
\item{darch}{a DArch instance}

\item{num_of_layer}{the number of the layer to print}

\item{input}{an input dataset. Defaults to darch's input data}

\item{show_derivative}{whether to show the activation or the derivative}

\item{type}{type of the graphs to print. It can be heatmap, surface, or histogram}
}
\description{
Prints out the network's layer activation value
}

