% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/util.R
\name{print_weight}
\alias{print_weight}
\title{Prints out the weight of a deep neural network}
\usage{
print_weight(darch, num_of_layer, show_derivative = F, type = "heatmap")
}
\arguments{
\item{darch}{DArch instance}

\item{num_of_layer}{the number of the layer to print}

\item{show_derivative}{T to show the weight value. F to show the percentage
weight change in the finetuning stage. This helps spot the network saturation problem.}

\item{type}{type of the graph. It supports "heatmap", "surface", and "histogram"}
}
\description{
This function prints out the weight in a heat map, 3D surface, or histogram
}

