% Generated by roxygen2 (4.0.0): do not edit by hand
\name{reconstruct}
\alias{reconstruct}
\alias{reconstruct.DeepBeliefNet}
\alias{reconstruct.RestrictedBolzmannMachine}
\title{Reconstruct data through a Deep Belief Nets and Restricted Bolzman Machines}
\usage{
reconstruct(object, newdata, ...)

\method{reconstruct}{DeepBeliefNet}(object, newdata, drop = TRUE, ...)

\method{reconstruct}{RestrictedBolzmannMachine}(object, newdata, drop = TRUE,
  ...)
}
\arguments{
\item{object}{the \code{\link{RestrictedBolzmannMachine}} or \code{\link{DeepBeliefNet}} object}

\item{newdata}{a \code{\link{data.frame}} or \code{\link{matrix}} providing the data. Must have the same columns than the input layer of the model.}

\item{drop}{do not return additional dimensions}

\item{\dots}{ignored}
}
\value{
the reconstructed data
}
\description{
Passes the data all the way through an unrolled DeepBeliefNet (in this case, it is identical to predict).
For a RestrictedBolzmannMachine or a DeepBeliefNet that hasn't been unrolled, it will predict, and predict again through the reversed network.
In the end, the reconstruction has the same dimension as the input.
}

