% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rectified_linear_unit.R
\name{rectified_linear_unit_function}
\alias{rectified_linear_unit_function}
\title{Rectified Linear Unit Function}
\usage{
rectified_linear_unit_function(data, weights)
}
\arguments{
\item{data}{the data matrix for calculation}

\item{weights}{the connection (weight matrix/filter) and the bias}
}
\value{
A list of function values and derivatives
}
\description{
This functions calculates the value and the derivative of a rectified linear
function. Reference Vinod Nair, Geoffrey Hinton, Rectified Linear Units
Improve Restricted Boltzmann Machines
}

