% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/resample.R
\name{resample}
\alias{resample}
\alias{resample.DeepBeliefNet}
\alias{resample.RestrictedBolzmannMachine}
\title{Sampling Methods for Deep Belief Nets and Restricted Bolzman Machines}
\usage{
resample(...)

\method{resample}{DeepBeliefNet}(object, newdata, drop = TRUE, ...)

\method{resample}{RestrictedBolzmannMachine}(object, newdata, drop = TRUE,
  ...)
}
\arguments{
\item{object}{the model}

\item{newdata}{a \code{\link{data.frame}} or \code{\link{matrix}} providing the data to sample. Must have the same columns than the input layer of the model.}

\item{drop}{do not return additional dimensions}

\item{\dots}{ignored}
}
\description{
Sample from a \code{\link{DeepBeliefNet}} or \code{\link{RestrictedBolzmannMachine}} object
}
\examples{
library(mnist)
data(mnist)
## Make predictions on a DBN object
data(pretrained.mnist)
res <- resample(pretrained.mnist, mnist$test$x)
head(res)
# Contrast with predict
pred <- predict(pretrained.mnist, mnist$test$x)
head(pred)

## Sample an RBM object
rbm <- pretrained.mnist[[1]]
res <- resample(rbm, mnist$test$x)
head(res)
}

