% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/reset_population_mu_delta.R
\name{reset_population_mu_sigma}
\alias{reset_population_mu_sigma}
\title{Resets the mu and sigmas of a darch instance to 0 and 1}
\usage{
reset_population_mu_sigma(darch)
}
\description{
Resets the mu and sigmas of a darch instance
}

