% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/dbn.methods.R, R/rbm.methods.R
\name{rev}
\alias{rev}
\alias{rev.DeepBeliefNet}
\alias{rev.RestrictedBolzmannMachine}
\title{Reverse a Deep Belief Net}
\usage{
\method{rev}{DeepBeliefNet}(x)

\method{rev}{RestrictedBolzmannMachine}(x)
}
\arguments{
\item{x}{the DeepBeliefNet to reverse}
}
\value{
the reversed DeepBeliefNet
}
\description{
\code{rev} returns a reversed \code{DeepBeliefNet} object. Precisely, the output is converted into the input of the network, and conversely.
}
\examples{
rbm <- RestrictedBolzmannMachine(Layer(784, "continuous"), Layer(1000, "gaussian"))
rev(rbm)
}
\seealso{
\code{\link{RestrictedBolzmannMachine}}, \code{\link{DeepBeliefNet}}, \code{\link{rev}}
}

