% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rsq.R
\name{rsq.lm}
\alias{rsq.lm}
\title{Utilitiy function that calcualtes RSQ of a linear model}
\usage{
\method{rsq}{lm}(x, input, target, ...)
}
\arguments{
\item{x}{linear Model}

\item{input}{Input data}

\item{target}{Target data}

\item{...}{additional inputs}
}
\description{
Calcualte a regression model's RSQ
}

