% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rsq.R
\name{rsq}
\alias{rsq}
\title{Calculate the RSQ of a regression model
Utilitiy function that calcualtes RSQ of a model. It measures the goodness-of-
fit of a regression model.}
\usage{
rsq(x, ...)
}
\arguments{
\item{x}{Regression Model}

\item{...}{Additional Input}
}
\description{
Calculate the RSQ of a regression model
Utilitiy function that calcualtes RSQ of a model. It measures the goodness-of-
fit of a regression model.
}

