% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/run_darch_bn.R
\name{run_darch_bn}
\alias{run_darch_bn}
\title{Execution function that runs in the batch normalization mode}
\usage{
run_darch_bn(darch, data)
}
\description{
Execution function that runs in the batch normalization mode
}

