% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/run_dnn.R
\name{run_dnn}
\alias{run_dnn}
\title{Execution function that runs in the batch normalization mode}
\usage{
run_dnn(darch, data)
}
\arguments{
\item{darch}{a darch instance}

\item{data}{input data}
}
\description{
This function calcualtes the output of a deep neural network with input data
}

