% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/train_dnn_bn.R
\name{train_dnn_bn}
\alias{train_dnn_bn}
\title{Training function with batch normalization on}
\usage{
train_dnn_bn(darch, input, target, num_epochs, learn_rate_gamma = exp(-10),
  error_function = meanSquareErr)
}
\description{
Train a deep neural network with batch normalization
}

