% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/util.R
\name{verticalize}
\alias{verticalize}
\title{Creates a matrix by repeating a row vector N times}
\usage{
verticalize(vector, N)
}
\arguments{
\item{vector}{the row vector}

\item{N}{number of rows in the output matirx}
}
\value{
a matrix
}
\description{
helper function that repeat a row vector N times
}

