% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/AR.R
\name{AR.DArch}
\alias{AR.DArch}
\title{Calculates the Accruacy Ratio of a given set of probability}
\usage{
\method{AR}{DArch}(darch, input = darch@dataSet@data,
  target = darch@dataSet@targets)
}
\arguments{
\item{target}{binary response}

\item{x}{a list of model output in the form of probabilities}
}
\description{
This function calculates the Accuracy Ratio of a trained darch instance
}

