% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/util.R
\name{AR.DArch}
\alias{AR.DArch}
\title{Utility function that calculates the Accuracy Ratio of a binary classification
 model}
\usage{
\method{AR}{DArch}(darch, input = darch@dataSet@data,
  target = darch@dataSet@targets)
}
\arguments{
\item{input}{input variable}

\item{target}{target variable}

\item{x}{an instance of DArch class}
}
\description{
Calculates the Accruacy Ratio of a given set of probability
}

