% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/AR.R
\name{AR.numeric}
\alias{AR.numeric}
\title{Calculates the Accruacy Ratio of a given set of probability}
\usage{
\method{AR}{numeric}(x, target)
}
\arguments{
\item{x}{a list of model output in the form of probabilities}

\item{target}{binary response}
}
\description{
This function calculates the Accuracy Ratio of a binary classification model
 output against its targets
}

