% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/backpropagate_delta_bn.R
\name{backpropagate_delta_bn}
\alias{backpropagate_delta_bn}
\title{function that calculates the delta function of a darch object with batch
normalization}
\usage{
backpropagate_delta_bn(darch, trainData, targetData,
  errorFunc = meanSquareErr)
}
\description{
Calculates the delta functions using backpropagation
}

