% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/calculate_population_mu_delta.R
\name{calcualte_population_mu_sigma}
\alias{calcualte_population_mu_sigma}
\title{Calculates the mu and sigmas of a darch instance}
\usage{
calcualte_population_mu_sigma(darch, input)
}
\description{
Calculates the mu and sigmas of a darch instance
}

