% Generated by roxygen2 (4.0.0): do not edit by hand
\name{continue.functions}
\alias{continue.function.always}
\alias{continue.function.exponential}
\alias{continue.function.exponential.aic}
\alias{continue.function.exponential.bic}
\alias{continue.function.random}
\alias{continue.functions}
\title{Continue functions}
\usage{
continue.function.exponential(error, iter, batchsize, ic = AIC)

continue.function.always(error, iter, batchsize)

continue.function.exponential.aic(error, iter, batchsize)

continue.function.exponential.bic(error, iter, batchsize)

continue.function.random(error, iter, batchsize)
}
\arguments{
\item{error}{a vector of the errors along the training}

\item{iter,batchsize}{current iteration number and batchsize.}
}
\value{
boolean (see description)
}
\description{
Functions to stop (by returning \code{FALSE}) the training if it has converged.

\code{continue.function.exponential} fits an exponential to the error and return \code{TRUE} if the function hasn't converged (or in case of doubt), \code{FALSE} if the function has converged or if the data couldn't be fitted (plateau reached or no exponential fit).
As a side effect, this function plots the error and the fit, and prints a summary of the fit on the console

\code{continue.function.always} always returns true so that the training carries on until \code{maxiters} is reached
}

