% Generated by roxygen2 (4.0.0): do not edit by hand
\name{drop}
\alias{drop}
\title{Drop a DeepBeliefNet}
\usage{
drop(x)
}
\arguments{
\item{x}{the DeepBeliefNet}
}
\description{
Drops the DeepBeliefNet information if it contains only a single RestrictedBolzmannMachine.
Provides an alias to \code{\link[base]{drop}} as \code{drop.default}.
}
\examples{
dbn <- dbn(layers=c(784, 1000), input="continuous", output="binary")
print(dbn)
rbm <- drop(dbn)
print(rbm)
}
\seealso{
\code{\link{DeepBeliefNet}}, \code{\link[base]{drop}}
}

