% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/energy.R
\name{energy}
\alias{energy}
\alias{energy.DeepBeliefNet}
\alias{energy.RestrictedBolzmannMachine}
\title{Energy of the Deep Belief Net or Restricted Bolzmann Machine.}
\usage{
energy(x, data, ...)

\method{energy}{RestrictedBolzmannMachine}(x, data, prediction, drop = TRUE,
  ...)

\method{energy}{DeepBeliefNet}(x, data, drop = TRUE, ...)
}
\arguments{
\item{x}{the \code{\link{DeepBeliefNet}} or \code{\link{RestrictedBolzmannMachine}} object}

\item{data}{the dataset, either as matrix or data.frame. The number of columns must match the number of nodes of the network input}

\item{prediction}{the predicted values of the hidden layer, as returned by \code{\link{predict}}. If provided (in case it was already calculated before), will speed-up the calculation.}

\item{drop}{do not return additional dimensions}

\item{\dots}{ignored}
}
\value{
a vector or matrix of the same size than the data (rows) giving the energy of each data point
}
\description{
Computes the energy of the data points in the DeepBeliefNet or RestrictedBolzmannMachine
}

