% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/energy.R
\name{energy}
\alias{energy}
\alias{energy.RestrictedBolzmannMachine}
\alias{energy.DeepBeliefNet}
\title{Energy of the Deep Belief Net or Restricted Bolzmann Machine.}
\usage{
energy(x, data, ...)

\method{energy}{RestrictedBolzmannMachine}(x, data, drop = TRUE, ...)

\method{energy}{DeepBeliefNet}(x, data, drop = TRUE, ...)
}
\arguments{
\item{x}{the \code{\link{DeepBeliefNet}} or \code{\link{RestrictedBolzmannMachine}} object}

\item{data}{the dataset, either as matrix or data.frame. The number of columns must match the number of nodes of the network input}

\item{\dots}{ignored}

\item{drop}{do not return additional dimensions}
}
\value{
a vector or matrix of the same size than the data (rows) giving the energy of each data point
}
\description{
Computes the energy of the data points in the DeepBeliefNet or RestrictedBolzmannMachine
}
\examples{
library(mnist)
data(mnist)

# Calculate error per data point on RBM
data(pretrained.mnist)
rbm <- pretrained.mnist[[1]]
en <- energy(rbm, mnist$test$x)
head(en) # 1 value per data point

# Calculate error per data point on DBN
data(trained.mnist)
en <- energy(trained.mnist, mnist$test$x)
head(en) # 1 value per data point

# Energy is not related with reconstruction error
err <- error(trained.mnist, mnist$test$x)
\dontrun{plot.mnist(predictions = cbind(err, en))}
}
