% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/error.R
\name{error}
\alias{error}
\alias{error.DeepBeliefNet}
\alias{rmse}
\alias{rmse.DeepBeliefNet}
\title{Compute Reconstruction Error}
\usage{
error(...)

rmse(...)

\method{error}{DeepBeliefNet}(x, data, ...)

\method{rmse}{DeepBeliefNet}(x, data, ...)
}
\arguments{
\item{...}{further arguments to the \code{plot} function above and to the \code{predict} function.}

\item{x}{the \code{\link{RestrictedBolzmannMachine}} or \code{\link{DeepBeliefNet}} object}

\item{data}{feature vector}
}
\value{
the reconstruction error of the data
}
\description{
Computes the reconstruction error (rmse) of the prediction of the data. \code{rmse} is an alias for \code{error}.
}

