% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/error.R, R/evidence.gradient.R
\name{error}
\alias{error}
\alias{error.DeepBeliefNet}
\alias{error.RestrictedBolzmannMachine}
\alias{errorSum}
\alias{errorSum.DeepBeliefNet}
\alias{errorSum.RestrictedBolzmannMachine}
\alias{evidence.gradient.RestrictedBolzmannMachine}
\alias{rmse}
\alias{rmse.DeepBeliefNet}
\alias{rmse.RestrictedBolzmannMachine}
\title{Compute Reconstruction Error}
\usage{
error(...)

rmse(...)

errorSum(...)

\method{error}{DeepBeliefNet}(x, data, ...)

\method{rmse}{DeepBeliefNet}(x, data, ...)

\method{errorSum}{DeepBeliefNet}(x, data, ...)

\method{error}{RestrictedBolzmannMachine}(x, data, ...)

\method{rmse}{RestrictedBolzmannMachine}(x, data, ...)

\method{errorSum}{RestrictedBolzmannMachine}(x, data, ...)

\method{evidence.gradient}{RestrictedBolzmannMachine}(x, data, ...)
}
\arguments{
\item{...}{further arguments to the \code{plot} function above and to the \code{predict} function.}

\item{x}{the \code{\link{RestrictedBolzmannMachine}} or \code{\link{DeepBeliefNet}} object}

\item{data}{feature vector}
}
\value{
the reconstruction error of the data
}
\description{
Computes the reconstruction error (rmse) of the prediction of the data. \code{rmse} is an alias for \code{error}. \code{errorSum} sum the error over the data points.
}

