% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/error.R
\name{error}
\alias{error}
\alias{rmse}
\alias{errorSum}
\alias{error.DeepBeliefNet}
\alias{rmse.DeepBeliefNet}
\alias{errorSum.DeepBeliefNet}
\alias{error.RestrictedBolzmannMachine}
\alias{rmse.RestrictedBolzmannMachine}
\alias{errorSum.RestrictedBolzmannMachine}
\title{Compute Reconstruction Error}
\usage{
error(...)

rmse(...)

errorSum(...)

\method{error}{DeepBeliefNet}(x, data, ...)

\method{rmse}{DeepBeliefNet}(x, data, ...)

\method{errorSum}{DeepBeliefNet}(x, data, ...)

\method{error}{RestrictedBolzmannMachine}(x, data, ...)

\method{rmse}{RestrictedBolzmannMachine}(x, data, ...)

\method{errorSum}{RestrictedBolzmannMachine}(x, data, ...)
}
\arguments{
\item{...}{further arguments to the \code{plot} function above and to the \code{predict} function.}

\item{x}{the \code{\link{RestrictedBolzmannMachine}} or \code{\link{DeepBeliefNet}} object}

\item{data}{feature vector}
}
\value{
the reconstruction error of the data
}
\description{
Computes the reconstruction error (rmse) of the prediction of the data. \code{rmse} is an alias for \code{error}. \code{errorSum} sum the error over the data points.
}
\examples{
library(mnist)
data(mnist)
data(trained.mnist)

# Calculate (reconstruction) error per data point
err <- error(trained.mnist, mnist$test$x)
length(err) # 1 value per data point
# error and rmse are synonymous
identical(err, rmse(trained.mnist, mnist$test$x))

# errorSum returns the sum
sum <- errorSum(trained.mnist, mnist$test$x)
print(sum)
all.equal(sum, sum(err)) 
# There may be some rounding errors though, so this might not be ==:
sum == sum(err)

# On a RestrictedBolzmannMachine
data(pretrained.mnist)
rbm <- pretrained.mnist[[1]]
errorSum(rbm, mnist$test$x)
err <- error(rbm, mnist$test$x)
identical(err, rmse(rbm, mnist$test$x))

}
