% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/finetune_SGD_bn.R
\name{finetune_SGD_bn}
\alias{finetune_SGD_bn}
\title{Finetune function that trains a deep neural network using stochastic gradient descent method}
\usage{
finetune_SGD_bn(darch, trainData, targetData, learn_rate_weight = exp(-10),
  learn_rate_bias = exp(-10), learn_rate_gamma = exp(-10),
  errorFunc = meanSquareErr, with_BN = T)
}
\description{
This function finetunes a DArch network using SGD approach
}

