% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/util.R
\name{forward_propagate}
\alias{forward_propagate}
\title{Forward propagate a deep neural network}
\usage{
forward_propagate(darch, input = darch@dataSet@data)
}
\arguments{
\item{darch}{A DArch instance}

\item{input}{Input data}
}
\description{
Calcualte all layer's activation value and derivative function value
}

