% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dropout.R
\name{generateDropoutMask}
\alias{generateDropoutMask}
\title{Generates the dropout mask for the deep neural network}
\usage{
generateDropoutMask(length, dropoutRate)
}
\description{
Generates the dropout mask for the deep neural network
}

