% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dropout.R
\name{generateDropoutMasksForDarch}
\alias{generateDropoutMasksForDarch}
\title{Generates dropout masks for dnn}
\usage{
generateDropoutMasksForDarch(darch, dropout_input, dropout_hidden)
}
\description{
Generates dropout maks for dnn
}

