% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/predict.R
\name{predict}
\alias{predict}
\alias{predict.DeepBeliefNet}
\alias{predict.RestrictedBolzmannMachine}
\title{Predict Methods for Deep Belief Nets and Restricted Bolzman Machines}
\usage{
\method{predict}{DeepBeliefNet}(object, newdata, drop = TRUE, ...)

\method{predict}{RestrictedBolzmannMachine}(object, newdata, drop = TRUE,
  ...)
}
\arguments{
\item{object}{the model}

\item{newdata}{a \code{\link{data.frame}} or \code{\link{matrix}} providing the data. Must have the same columns than the input layer of the model.}

\item{drop}{do not return additional dimensions}

\item{\dots}{ignored}
}
\description{
Obtain predictions from a \code{\link{DeepBeliefNet}} or \code{\link{RestrictedBolzmannMachine}} object
}
\examples{
library(mnist)
data(mnist)
## Make predictions on a DBN object
data(trained.mnist)
predict(trained.mnist, mnist$test$x[1:10,])
## Make predictions on a RBM object
data(pretrained.mnist)
rbm <- pretrained.mnist[[1]]
predictions <- predict(rbm, mnist$test$x)
dim(predictions) # 1000 columns, output size of the rbm
ncol(predictions) == rbm$output$size
}
