% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/calculate_mu_sigma.R
\name{reset_population_mu_sigma}
\alias{reset_population_mu_sigma}
\title{Resets the mu and sigmas of a darch instance to 0 and 1}
\usage{
reset_population_mu_sigma(darch)
}
\arguments{
\item{darch}{a darch instance}
}
\description{
This function resets the mu and sigmas of hidden layers in a darch instance
 to 0 and 1
}

