% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rsq.R
\name{rsq.DArch}
\alias{rsq.DArch}
\title{Utilitiy function that calcualtes RSQ of a DArch instance}
\usage{
\method{rsq}{DArch}(darch, input = darch@dataSet@data,
  target = darch@dataSet@targets)
}
\arguments{
\item{input}{Input data}

\item{target}{Target data}

\item{x}{DArch Model}
}
\description{
Calcualte a regression model's RSQ of a deep neural network
}

