% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/util.R
\name{rsq.lm}
\alias{rsq.lm}
\title{Utilitiy function that calcualtes RSQ of a linear model}
\usage{
\method{rsq}{lm}(mod, input, target)
}
\arguments{
\item{input}{Input data}

\item{target}{Target data}

\item{x}{linear Model}
}
\description{
Calcualte a regression model's RSQ
}

