% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/train_dnn.R
\name{train_dnn}
\alias{train_dnn}
\title{Train a deep neural network}
\usage{
train_dnn(darch, input, target, input_valid = NULL, target_valid = NULL,
  ..., learn_rate_weight, learn_rate_bias, learn_rate_gamma, batch_size,
  batch_normalization, dropout_input, dropout_hidden, momentunm_initial,
  momentum_final, momentum_switch, num_epochs, error_function,
  report_classification_error = F)
}
\description{
Train a deep neural network
}

