% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/unroll.R
\name{unroll}
\alias{unroll}
\title{Unroll the Deep Belief Net}
\usage{
unroll(x)
}
\arguments{
\item{x}{the dbn object}
}
\value{
the unrolled dbn with the \code{unrolled} switch is set to \code{TRUE}.
}
\description{
Unrolling the DBN means stacking a reversed copy of itself. This create a (pre-trained) auto-encoder that can be
}
\examples{
data(pretrained.mnist)
unrolled <- unroll(pretrained.mnist)
}

